Conditional NPC talk & entry point
Edit of npcs (see scientist for an example) :
Discourse can now be circonstancial. One can specify a condition under the form of a parameterless function for each sequence of sentences. It should still be working for existing npc, but others can adapt their talk.
 * The form, which is not tested is an array of objects. Each object (except the first, default discussion) has a "condition" property (parameterless function) which decides when to use this discussion and a "text" property, which contain the array of lines of the onversation.
 * The variable made available to the functions is the game itself.

Added "entry point" :
An entry point is used to get into the game through encoded urls (used sha1 by caligatio http://caligatio.github.com/jsSHA) to identify them. This can be used to establish connexion with the outside world, might it be on the Web (achievement obtained through activity on another site, section of the site) or in the real world (QRCode found on an item).
Can be tested with

* Corrected hasty commit (changed place where the Entry Point is called)
* Added an example : ./client/?entrance=almostFoxy
* hastily defined *name* as the name of the user in npc dialogs

Almost forgot to reinstate rick…