Add form validation layout and functionality
This adds the ability to highlight invalid fields, as well as include validation error messages in the login / create new character forms.

This also slightly adjusts the behavior when the Enter key is pressed (rather than simply switching focus, it will actually attempt to log in, and will print an error for the first missing field). If the focus is already set to the first missing field, this makes the Enter key more responsive (previously, pressing Enter would have seemingly done nothing in this case).

This also fixes #119 (while I was adjusting the CSS to make space for the validation error messages, I also made sure that the "Create" link fits within the container).