Avoid running tryStartingGame again if we're already in the process of launching the game
This can happen if, for example, you hit enter more than once while logging in. The only visible impact seemed to be that they play button text (either "Login" or "Create") did not get restored properly (it stayed as "Loading"), but this has potential to cause other problems as well, and should be fixed.